+++ integrations/consensus/src/test/java/org/pragmatica/consensus/rabia/RabiaConsensusIntegrationTest.java
@@ -467,6 +467,11 @@ class RabiaConsensusIntegrationTest {
         public TimeSpan pingInterval() {
             return timeSpan(1).seconds();
         }
+
+        @Override
+        public TimeSpan helloTimeout() {
+            return timeSpan(5).seconds();
+        }
     }
 
     static class TestStateMachine implements StateMachine<TestCommand> {